Fix UI report directory and prize_pot type validation
- Mount correct engine report directory (data/reports) in docker-compose.prod.yml
- Fix local fallback path in game_engine_containerized.py for consistency
- Change prize_pot from int to float in GameSummary model (engine stores floats)

The UI was previously mounting /reports/ which contained corrupted log-parser
generated files instead of /data/reports/ where the game engine saves proper
reports with correct player IDs and gameplay data.

+++ traitorsim-ui/backend/app/routers/games.py
@@ -36,7 +36,7 @@ class GameSummary(BaseModel):
     name: str
     created_at: str
     total_days: int
-    prize_pot: int
+    prize_pot: float  # Engine stores as float
     winner: str
     rule_variant: str
     config_total_players: Optional[int] = None
